Remove 'const' from void* serviceData
serviceData is used as an opaque 'user data' handle to be passed back to the
TRviCallback that the user registers in rviRegisterService. rviRegisterService
takes a 'void*', but the internal 'rviServiceCreate' is defined to take a
'const void*'. I believe the right interface here is non-const, since
TRviCallback is free to change the data blob it gets via the 'void *
serviceData' parameter.